Now deals property with repated files

+++ pics_functions.sh
@@ -1,6 +1,11 @@
 rename_pic_if_exists() {
-    # Accept parameter $file
-    # $file is the name you want o use to save the file as
+    # Requires 2 parameters:
+    # $file         <== The name of the file bo backup
+    # $THEFOLDER    <== Where the file will be backedup
+    #
+    # If the file name is already present in the folder, an md5 checksum is done,
+    # if they are different, the existing file is renamed.
+    # If the sums match, nothing is done.
 
     #set -x
 
@@ -29,19 +34,25 @@ rename_pic_if_exists() {
     existing_file="$folder/$fbase"
 
     if [ -f "$existing_file" ]; then
-        # File exists, rename it
-        whole_string=$( exif "$existing_file" | grep 'Date and Time' | grep 'Origi' )
-        the_date=$( echo "$whole_string" | awk '{split($0,arr,"|"); print arr[2]}' | awk -F'[: ]' '{print $1$2$3$4$5$6}' )
-        the_date="$the_date"
+        # File exists, rename it only if md5 checksums are different, otherwise just skip over it
+        sum_existing_file=$( md5sum $existing_file | awk '{print $1}' )
+        sum_file_to_backup=$( md5sum $file_to_backup | awk '{print $1}' )
 
-        echo "prefix:"
-        echo "$fprefix"
+        if [ "$sum_existing_file" != "$sum_file_to_backup" ]; then
+            whole_string=$( exif "$existing_file" | grep 'Date and Time' | grep 'Origi' )
+            the_date=$( echo "$whole_string" | awk '{split($0,arr,"|"); print arr[2]}' | awk -F'[: ]' '{print $1$2$3$4$5$6}' )
+            the_date="$the_date"
 
-        new_file_name="$folder/$fprefix-$the_date.$fext"
+            echo "prefix:"
+            echo "$fprefix"
 
-        echo "File exists: $existing_file"
-        echo "Renaming to $new_file_name"
+            new_file_name="$folder/$fprefix-$the_date.$fext"
 
-        mv "$existing_file" "$new_file_name"
+            echo "A File with the same name but different md5 sum exists."
+            echo "Renaming existing file: $existing_file"
+            echo "Renaming to $new_file_name"
+
+            mv "$existing_file" "$new_file_name"
+        fi
     fi
 }
